==> study/Level1-CellClass/ALL_manual/7-CellBrowser.ipynb <==
<h1>Table of Contents<span class="tocSkip"></span></h1>
<div class="toc"><ul class="toc-item"><li><span><a href="#Path" data-toc-modified-id="Path-1"><span class="toc-item-num">1&nbsp;&nbsp;</span>Path</a></span></li><li><span><a href="#Gene-Meta" data-toc-modified-id="Gene-Meta-2"><span class="toc-item-num">2&nbsp;&nbsp;</span>Gene Meta</a></span></li><li><span><a href="#Cell-Meta" data-toc-modified-id="Cell-Meta-3"><span class="toc-item-num">3&nbsp;&nbsp;</span>Cell Meta</a></span><ul class="toc-item"><li><span><a href="#Determine-meta-cols" data-toc-modified-id="Determine-meta-cols-3.1"><span class="toc-item-num">3.1&nbsp;&nbsp;</span>Determine meta cols</a></span></li><li><span><a href="#Select-meta-cols" data-toc-modified-id="Select-meta-cols-3.2"><span class="toc-item-num">3.2&nbsp;&nbsp;</span>Select meta cols</a></span></li></ul></li><li><span><a href="#Coords" data-toc-modified-id="Coords-4"><span class="toc-item-num">4&nbsp;&nbsp;</span>Coords</a></span></li><li><span><a href="#Gene-Matrix" data-toc-modified-id="Gene-Matrix-5"><span class="toc-item-num">5&nbsp;&nbsp;</span>Gene Matrix</a></span><ul class="toc-item"><li><span><a href="#Determine-genes-to-include" data-toc-modified-id="Determine-genes-to-include-5.1"><span class="toc-item-num">5.1&nbsp;&nbsp;</span>Determine genes to include</a></span></li><li><span><a href="#Save-gene-matrix" data-toc-modified-id="Save-gene-matrix-5.2"><span class="toc-item-num">5.2&nbsp;&nbsp;</span>Save gene matrix</a></span></li></ul></li><li><span><a href="#Conf-file" data-toc-modified-id="Conf-file-6"><span class="toc-item-num">6&nbsp;&nbsp;</span>Conf file</a></span></li><li><span><a href="#Color-File" data-toc-modified-id="Color-File-7"><span class="toc-item-num">7&nbsp;&nbsp;</span>Color File</a></span></li></ul></div>

import pathlib
import pandas as pd

import seaborn as sns
import matplotlib.pyplot as plt

result_dir = pathlib.Path('CellBrowser').absolute()
result_dir.mkdir(exist_ok=True)

## Path

page_name = 'CEMBA_Test'

cell_meta_path = './cell_tidy_data.final_annotation.with_region.msg'

gene_mcds_path_pattern = './Adata/SelectedCell*gene_da_rate*mcds'
gene_annotation_path = '/home/hanliu/ref/mouse/gencode/vm22/gencode.vM22.annotation.gene.flat.tsv.gz'

goi_path = None
MUST_INCLUDE_GENE_TYPE = ['protein_coding']

mc_types = ['CHN']

CELL_META_PATH = result_dir / f'{page_name}.cell_meta.tsv'

default_cluster_col = 'CellClass'
default_label_col = 'CellClass'

## Gene Meta

gene_meta = pd.read_csv(gene_annotation_path, index_col='gene_id', sep='\t')
gene_meta.index.name = 'gene'
gene_name_to_id = {v:k for k, v in gene_meta['gene_name'].iteritems()}
gene_id_base_to_id = pd.Series(gene_meta.index, index = gene_meta.index.map(lambda i: i.split('.')[0])).to_dict()

## Cell Meta

cell_meta = pd.read_msgpack(cell_meta_path)
print(cell_meta.shape)
print(cell_meta.columns.tolist())
cell_meta.head()

### Determine meta cols

exclude_cols = ['AllcPath']

categorical_cols = ['Row384', 'Col384', 'Row96', 'Col96', 'Pos96', 
                    'FACS_Date', 'PassFilter', 'Plate', 'Region', 
                    'Slice', 'index_name', 'uid', 'leiden', 
                    'CellClass', 'MajorRegion', 'SubRegion']

### Select meta cols

# save the non-coords part
cell_meta_no_coords = cell_meta.loc[:, (~(cell_meta.columns.str.endswith('_0') | 
                                         cell_meta.columns.str.endswith('_1'))) & 
                                       (~cell_meta.columns.isin(exclude_cols))].copy()

cell_meta_no_coords.index.name = 'Cell'

CELL_META_PATH = result_dir / f'{page_name}.cell_meta.tsv'
cell_meta_no_coords.to_csv(CELL_META_PATH, sep='\t')

## Coords

# save the non-coords part
cell_meta_coords = cell_meta.loc[:, (cell_meta.columns.str.endswith('_0') | 
                                     cell_meta.columns.str.endswith('_1'))].copy()

COORD_PATH_LIST = []

for coord_base, coord_df in cell_meta_coords.groupby(cell_meta_coords.columns.str[:-2], axis=1):
    print(coord_base.upper(), 'included in browser')
    # cellBrowser only support 2D
    coord_df = coord_df[[f'{coord_base}_0', f'{coord_base}_1']].copy()
    coord_df.index.name = 'cellId'
    coord_df.columns = ['x', 'y']
    
    coord_path = result_dir / f'{page_name}.{coord_base}.tsv'
    coord_df.to_csv(coord_path, sep='\t')
    COORD_PATH_LIST.append({'file': str(coord_path), 
                            'shortLabel': coord_base.upper()})
COORD_PATH_LIST = COORD_PATH_LIST[::-1]

## Gene Matrix

from ALLCools.mcds.MCDS import MCDS

gene_mcds = MCDS.open(gene_mcds_path_pattern,
                      combine='nested',
                      concat_dim='cell').sel(mc_type=mc_types, cell=cell_meta.index)

gene_mcds

### Determine genes to include

#### Curated GOI

if goi_path is not None:
    goi = pd.read_csv(goi_path, 
                      index_col=0, header=None).index
else:
    goi = pd.Index([])

#### Cell Type Markers

markers = pd.Index([])
for mc_type in mc_types:
    try:
        _markers = pd.read_csv(f'./Marker/{mc_type}_selected_feature.txt', 
                               index_col=0, header=None).index
        markers.append(_markers)
    except FileNotFoundError:
        pass

#### Must include gene types

judge = gene_meta['gene_type'].isin(MUST_INCLUDE_GENE_TYPE)
must_include_genes = gene_meta.index[judge]

#### Total genes

total_genes = must_include_genes | goi | markers
print(total_genes.size, 'genes to include in browser')

### Save gene matrix

GENE_PATH_DICT = {}

for mc_type in mc_types:
    _this_da = gene_mcds.sel(gene=total_genes).squeeze()['gene_da']
    # gene_df should be gene by cell, row is gene
    gene_df = _this_da.to_pandas().T
    gene_df.columns.name = ''
    gene_df.index = gene_df.index.map(lambda i: f'{i}|{gene_meta.loc[i, "gene_name"]}')
    gene_df_path = result_dir / f'{page_name}.{mc_type}.gene.tsv.gz'
    gene_df.to_csv(gene_df_path, sep='\t', float_format='%.3f')
    GENE_PATH_DICT[mc_type] = gene_df_path
                                      
    del gene_df
del gene_mcds

## Conf file

for mc_type in mc_types:
    default_config_dict = {
        'page_name': page_name,
        'page_priority': 10,
        'page_tags': [],
        'page_short_label': page_name,
        'gene_matrix_path': pathlib.Path(GENE_PATH_DICT[mc_type]).absolute(),
        'cell_meta_path': CELL_META_PATH,
        'categorical_cols': categorical_cols,
        'coord_file_list': COORD_PATH_LIST,
        'default_cluster_field': default_cluster_col,
        'default_label_field': default_label_col,
        'gene_value_unit': f'Norm {mc_type}%',
    }

with open('./CellBrowser/cellbrowser_template.conf') as f, open(result_dir / 'cellbrowser.conf', 'w') as out_f:
    config_str_template = f.read()
    config_str = config_str_template.format(**default_config_dict)
    out_f.write(config_str)

print(config_str)

## Color File
